// src/types/contact.ts
export interface Contact {
  id: number;
  phoneNumber?: string;
  email?: string;
  linkedId?: number;
  linkPrecedence: 'primary' | 'secondary';
  createdAt: Date;
  updatedAt: Date;
  deletedAt?: Date;
}

export interface ContactRequest {
  email?: string;
  phoneNumber?: string;
}

export interface ContactResponse {
  contact: {
    primaryContatctId: number;
    emails: string[];
    phoneNumbers: string[];
    secondaryContactIds: number[];
  };
}

export interface DatabaseContact {
  id: number;
  phone_number?: string;
  email?: string;
  linked_id?: number;
  link_precedence: 'primary' | 'secondary';
  created_at: Date;
  updated_at: Date;
  deleted_at?: Date;
}